Remove existing records from handler and update readme.md
To simplify the cloud kit handler protocol the field exisitingRecords was removed. This means parseResults only will receive the records fetched during that batch, so they have to be appended to the existing data.

This commit also updates readme with documentation for cloud kit error handler, cloud kit fetcher and cloud kit handler.